backend/requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
requests==2.25.1
oauthlib==3.1.1
pyOpenSSL==20.0.1
Flask-Login==0.5.0
flask-cors==3.0.10
python-dotenv==0.18.0